Rename sort to sorting
